Changed functioncalls to be parsed as Accessors.
This means we can now parse something like: object.field[12](argument);
instead of the call[closure)(args) syntax.
For the mean time this also means that most normal calls
do a global lookup for a closure name. Also, the nonsenical
foo(a,b) = 20; now interprets but mangles the stack.

Both these issues will be fixed when the AST introspection
is improved.